move the write objects of fetch to the scope-component-importer, change importMany to alwasys persist
