+++ src/scope/component-ops/scope-components-importer.ts
@@ -1,6 +1,7 @@
 import { filter } from 'bluebird';
 import { Mutex } from 'async-mutex';
 import mapSeries from 'p-map-series';
+import pMap from 'p-map';
 import groupArray from 'group-array';
 import R from 'ramda';
 import { compact, flatten } from 'lodash';
@@ -20,11 +21,12 @@ import ComponentVersion, { ObjectCollector } from '../component-version';
 import { ComponentNotFound } from '../exceptions';
 import { Lane, ModelComponent, Version } from '../models';
 import { Ref, Repository } from '../objects';
-import { ObjectItem } from '../objects/object-list';
+import { ObjectItem, ObjectList } from '../objects/object-list';
 import SourcesRepository, { ComponentDef } from '../repositories/sources';
 import { getScopeRemotes } from '../scope-remotes';
 import VersionDependencies from '../version-dependencies';
-import { DEFAULT_LANE } from '../../constants';
+import { CONCURRENT_COMPONENTS_LIMIT, DEFAULT_LANE } from '../../constants';
+import { BitObjectList } from '../objects/bit-object-list';
 
 const removeNils = R.reject(R.isNil);
 
@@ -64,12 +66,7 @@ export default class ScopeComponentsImporter {
    * 3. b. If all flattened exists locally - exit the loop.
    * 3. c. otherwise, put it in the externalsToFetch array.
    */
-  async importMany(
-    ids: BitIds,
-    cache = true,
-    persist = true,
-    throwForDependencyNotFound = false
-  ): Promise<VersionDependencies[]> {
+  async importMany(ids: BitIds, cache = true, throwForDependencyNotFound = false): Promise<VersionDependencies[]> {
     logger.debugAndAddBreadCrumb(
       'importMany',
       `cache ${cache}, throwForDependencyNotFound: ${throwForDependencyNotFound}. ids: {ids}`,
@@ -98,7 +95,7 @@ export default class ScopeComponentsImporter {
     logger.debug('importMany', `total missing externals: ${uniqExternals.length}`);
     const remotes = await getScopeRemotes(this.scope);
     // we don't care about the VersionDeps returned here as it may belong to the dependencies
-    await this.getExternalMany(uniqExternals, remotes, persist, throwForDependencyNotFound);
+    await this.getExternalMany(uniqExternals, remotes, throwForDependencyNotFound);
     const versionDeps = await this.bitIdsToVersionDeps(idsToImport);
     logger.debug('importMany, completed!');
     return versionDeps;
@@ -250,7 +247,7 @@ export default class ScopeComponentsImporter {
     });
     loader.start(`got ${objectList.count()} objects from the remotes, merging them and writing to the filesystem`);
     logger.debugAndAddBreadCrumb('importManyDeltaWithoutDeps', 'writing them to the model');
-    await this.scope.writeManyObjectListToModel(objectListPerRemote, true, idsToFetch);
+    await this.writeManyObjectListToModel(objectListPerRemote, idsToFetch);
   }
 
   async importFromLanes(remoteLaneIds: RemoteLaneId[]): Promise<Lane[]> {
@@ -395,6 +392,51 @@ export default class ScopeComponentsImporter {
     return this.loadRemoteComponent(id);
   }
 
+  async writeManyObjectListToModel(
+    objectListPerRemote: { [remoteName: string]: ObjectList },
+    ids: BitId[]
+  ): Promise<BitId[]> {
+    const bitObjectsPerRemote: { [remoteName: string]: BitObjectList } = {};
+    await mapSeries(Object.keys(objectListPerRemote), async (remoteName) => {
+      const objectList = objectListPerRemote[remoteName];
+      bitObjectsPerRemote[remoteName] = await objectList.toBitObjects();
+    });
+    const bitIds = await mapSeries(Object.keys(bitObjectsPerRemote), async (remoteName) => {
+      const bitObjectList = bitObjectsPerRemote[remoteName];
+      return this.addObjectListToRepo(bitObjectList, remoteName, ids);
+    });
+    await this.repo.persist();
+    return BitIds.uniqFromArray(R.flatten(bitIds));
+  }
+
+  private async addObjectListToRepo(bitObjectList: BitObjectList, remoteName: string, ids: BitId[]): Promise<BitId[]> {
+    const components = bitObjectList.getComponents();
+    const versions = bitObjectList.getVersions();
+    const laneObjects = bitObjectList.getLanes();
+    await pMap(
+      components,
+      (component: ModelComponent) => this.scope.mergeModelComponent(component, versions, remoteName),
+      {
+        concurrency: CONCURRENT_COMPONENTS_LIMIT,
+      }
+    );
+    let nonLaneIds: BitId[] = ids;
+    await Promise.all(
+      laneObjects.map(async (lane) => {
+        if (!lane.scope) {
+          throw new Error(`scope.addObjectListToRepo scope is missing from a lane ${lane.name}`);
+        }
+        await this.repo.remoteLanes.syncWithLaneObject(lane.scope, lane);
+        nonLaneIds = nonLaneIds.filter((id) => id.name !== lane.name || id.scope !== lane.scope);
+        nonLaneIds.push(...lane.components.map((c) => c.id));
+      })
+    );
+    // components and lanes were merged previously, add the rest.
+    const objectsToAdd = bitObjectList.getAllExceptComponentsAndLanes();
+    this.sources.putObjects(objectsToAdd);
+    return nonLaneIds;
+  }
+
   private async getVersionFromComponentDef(component: ModelComponent, id: BitId): Promise<Version | null> {
     const versionComp: ComponentVersion = component.toComponentVersion(id.version);
     const version = await versionComp.getVersion(this.scope.objects, false);
@@ -415,7 +457,6 @@ export default class ScopeComponentsImporter {
   private async getExternalMany(
     ids: BitId[],
     remotes: Remotes,
-    persist = true,
     throwForDependencyNotFound = false
   ): Promise<VersionDependencies[]> {
     if (!ids.length) return [];
@@ -435,7 +476,7 @@ export default class ScopeComponentsImporter {
       context
     );
     logger.debugAndAddBreadCrumb('ScopeComponentsImporter.getExternalMany', 'writing them to the model');
-    const nonLaneIds = await this.scope.writeManyObjectListToModel(objectListPerRemote, persist, ids);
+    const nonLaneIds = await this.writeManyObjectListToModel(objectListPerRemote, ids);
     const componentDefs = await this.sources.getMany(nonLaneIds);
     const componentDefsExisting = componentDefs.filter((componentDef) => componentDef.component);
     const versionDeps = await mapSeries(componentDefsExisting, (compDef) =>
@@ -473,7 +514,7 @@ export default class ScopeComponentsImporter {
       { withoutDependencies: true },
       context
     );
-    const nonLaneIds = await this.scope.writeManyObjectListToModel(objectListPerRemote, true, ids);
+    const nonLaneIds = await this.writeManyObjectListToModel(objectListPerRemote, ids);
 
     const finalDefs: ComponentDef[] = await this.sources.getMany(nonLaneIds);
